Ex3
Did Ex3
There are two version: Once where the State-Transition Diagram was  a Mealy-Automaton (SequenceDetector).
And another one, where it was a Moore-Automaton (SequenceDetectorMoore and _2 versions), leading to using all possible states (4) of the two Flip-FLops.

There is a problem with initializing the second one, because Logisim, can´t set default states of custom Flip-Flops